Rename audit table
