--- api/alembic/versions/353773d17996_adds_audit_model.py
@@ -1,41 +1,0 @@
-"""Adds Audit model
-
-Revision ID: 353773d17996
-Revises: 004ac48ffe18
-Create Date: 2021-04-13 08:31:58.014501
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '353773d17996'
-down_revision = '004ac48ffe18'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic ###
-    op.create_table('audits',
-                    sa.Column('id', sa.Integer(), nullable=False),
-                    sa.Column('create_user', sa.String(), nullable=False),
-                    sa.Column('path', sa.String(), nullable=False),
-                    sa.Column('create_timestamp', sa.TIMESTAMP(timezone=True), nullable=False),
-                    sa.PrimaryKeyConstraint('id'),
-                    comment='The audit log of an authenticated request by a user.'
-                    )
-    op.create_index(op.f('ix_audits_create_timestamp'), 'audits', ['create_timestamp'], unique=False)
-    op.create_index(op.f('ix_audits_create_user'), 'audits', ['create_user'], unique=False)
-    op.create_index(op.f('ix_audits_path'), 'audits', ['path'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_audits_path'), table_name='audits')
-    op.drop_index(op.f('ix_audits_create_user'), table_name='audits')
-    op.drop_index(op.f('ix_audits_create_timestamp'), table_name='audits')
-    op.drop_table('audits')
-    # ### end Alembic commands ###
